Add rule to forbid access to plugin.configs from src/configs
This is a custom rule that is admittedly very tailored to the plugin and so it is not included in the plugin base config rules. The issue this solves is that if you try to access plugin.configs from a file in src/configs, there's no guarantee that the plugin will have that given config at that point in time, since src/index.ts creates the config object with the helpers, which take in the plugin as a whole while other rules may still be getting created.

As such, it is safer to just recreate the config for the plugin by using one of the config creator functions instead. That way, we can guarantee that the config will exist.