Add arrow-body-style rule to allow enforce curly brackets around arrow functions
I tend to prefer always using curly brackets for arrow functions because it's easier to extend should you need to add to them. Any function that has more than one line to it will always need curly brackets, so any time you want to extend a one-line function, you have to add curly brackets around them every time, which can be fiddly, especially if the return is a very long object.

Furthermore, if you want to put a quick console.log before the return for debugging purposes to see when it executes, for example, this is less ideal without curly brackets because you have to insert them in, then write the console.log statement, and if, for whatever reason, we want to enforce no curly brackets, we then have to remove them again. As such, I've decided that it'd be best for them to just always be there.