Loosen up the configs type in plugin interface
This makes it easier to add new configs in the future, so we don't need to also add it to the interface as an extra step. The precise config names aren't super important so it's not worth specifying in the interface. What's more important is the structure of the configs - they need to be an object with values being config arrays, but the keys can be anything.