Allow depth of relative imports to be configured
The fixer for this rule never seemed to work properly, so it has been scrapped in favour of allowing a certain depth level of relative import instead. This allows my rule to stand out against the no-restricted-imports approach even though I can't quite provide the fixable solution (at least, not yet). That way, people who are at least fine with ./ can still have it while only disabling ../, and people who are fine with ../ but only up to a certain point can configure it that way too.