Add no-skipped-tests rule that warns instead of errors
Temporary skipped tests are often run as part of a pre-commit hook when committing changes that doesn't affect a currently failing test, so it's better to let that warn instead of throw an error so pre-commits can pass. That way, we're made aware of the potentially unintentional skipped test but it doesn't block us if we don't need it to.