Refactor to use new config setup
- All general-purpose configs that do not use plugin rules are now accessed using plugin.configs['general/config-name'].
- All configs that only use my plugin rules are now accessed using plugin.configs['plugin/config-name'].
- All configs that combine general rules and plugin rules are now accessed using plugin.configs['combined/config-name'].

This helps make it clearer exactly which rules are plugin-specific, which ones are general-purpose, and which ones are a combination of both. That way, people have more control over whether to use plugin-only rules, non-plugin rules, or default combinations. This should also make the configs more stable as a whole, given that we can isolate them better now so that the general config changes do not affect plugin-only configs and vice-versa.