Add standardise-error-messages rule
This ensures that all error messages passed into Error follow the same format. In the default case, this will be SOMETHING_LIKE_THIS. This allows the code to throw standardised errors that can be easier to check and keep track of, while something like the Neurosongs front-end can convert it into a human-readable string.