Get rid of Linter.Config[] type casting
This isn't really needed anymore since we're no longer directly importing @typescript-eslint/eslint-plugin, which was the main cause of this. Might be because that dependency was using an older version of ESLint and so the types didn't quite align, but it seems that using typescript-eslint and configuring the base TypeScript config with that gets rid of the need to do this.