Add personal config group
This helps us to separate out my specific personal preference rules so that they can then be more freely combined in the combined configs.

Note that I have also had to remove package.json linting from the general JavaScript configs because that tended to mess with the combination of the TypeScript configs. Package.json linting should most likely be its own config set anyway, but I'm not sure how worth it it would be as of now given that I mainly just use the recommended configs from packageJson anyway.