Type the plugin configs better to allow for autocompletion
Previously, autocomplete on the config property did not happen due to the string/string typing. This change now refactors the AlexPluginConfigs type such that the keys are extracted away into their own unions, and then we also have a ConfigKey type that knows how to resolve the overall flat type based on the format.