Restrict direct imports of the plugin in config files
This should never really be allowed as it can potentially create circular imports where we try and access the plugin from the configs while it's also trying to define its own configs. The better pattern is to define a function that takes in the current plugin's state, and then returns a config based on that. That way, we never have to reference the plugin directly from the configs themselves.