Abstract out language options
The language options tends to be fairly lengthy compared to the rest of the config, and also tends to be repeated across configs (e.g. the TypeScript language options are the exact same across all TypeScript configs). As such, it makes sense to separate them out so that they can be used across configs.