Use Zod to generate the options schema directly
Being able to use Zod in these schemas is incredibly helpful as it allows us to easily validate the options. Lucky for us, Zod has a nice way to convert its own Zod schemas into JSON schemas, which are usable in ESLint rules. Unfortunately, however, it seems the JSONSchema type from Zod is not compatible with the JSONSchema4 type from ESLint for some reason, and so a type cast is needed.