Update peer dependencies
Note that the separate @typescript-eslint/eslint-plugin and @typescript-eslint/parser dependencies had to be uninstalled as they were causing version conflicts. The base typescript-eslint package already provides both, so it's better to just use that as it's far easier to maintain one stable dependency rather than two unstable ones.

Also note that the eslint-plugin-react-hooks plugin seems to not be typed exactly how ESLint expects it to be. As of now, at least, it exports its flat configs on plugin.configs.flat.recommended, which is not quite what the built-in Linter.Config[] expects. As such, I have had to manually map the plugin to the correct type myself by flattening out plugin.configs.flat.recommended to plugin.configs['flat/recommended']. This seems to be the more conventional ESLint way to export grouped configs, and it is weird as to why eslint-plugin-react-hooks does not do this.