Uninstall eslint-plugin-eslint-plugin
I just tried using the recommended rules, and it broke literally everything. Not just in the sense of adding many suggestions to fix (which would be acceptable), but running the fixer literally just produced a ton of its own syntax errors... Not amazing for a plugin whose literal job is to lint plugins themselves.