Loosen peer dependency versions and specify stricter versions of peer dependencies in devDependencies
This is generally the better pattern to follow so that dev dependencies specify exactly which version ranges we are using, whereas peer dependencies specify which versions the package is currently compatible with. Peer dependencies should be loose so that we can allow for a wider range of compatible package versions per peer dependency, but they should be specified as dev dependencies as well so that we can still update them in this project.