Switch the no-namespace-imports tests to use eslint-vitest-rule-tester
I think this is better than the previous usage of @typescript-eslint/rule-tester directly as this allows us to structure the test suite in the classic describe, test, expect way rather than with a giant ruleTester.run invocation that is harder to control. This approach opens up the possibility of us being able to use something like tempy for ruleTesterWithParser tests where we may need file directory information.