Freeze the plugin when creating configs and type the plugin in config functions as Readonly
Config functions should never mutate the plugin directly - the whole config setup system depends on the fact that we're passing through a single instance of the plugin to all config creators so they can reference it but never mutate it directly. That way, the configs can reference that single plugin instance and use it in the config, and then once all configs have been created, they then get assigned to the final plugin. This change ensures that the plugin can never be accidentally mutated.

It is now typed as Readonly, so now the editor will error on mutation attempt, but even if we forget to do this, it has been frozen at runtime so it will error if a mutation attempt happens. Also note the use of deepFreeze instead of Object.freeze. This is because the plugin definitely is a heavily nested object and we need to ensure that each nested property is also frozen.

Also note the fact that we need to spread alexPlugin into deepFreeze. This is because we don't want to freeze the plugin itself as we still need to mutate it - we only want to freeze it when it is passed into the configs. That way, the configs can never mutate the plugin, but we can still do so in src/alexPlugin.ts to assign the configs to the plugin after they get created.