Move the no-restricted-imports to affect all files in the plugin repository
The plugin should never really be imported directly in any file as we can never really guarantee the current state of the plugin throughout this repository. This is true for the configs, but it could potentially be true for most other files too - the imported plugin here may not always be exactly the same as the final one, and it is often better to get the content you need by importing a specific part instead.