Deeply freeze a deeply-copied plugin before passing to configs
Before, I was only passing in a shallowly-copied plugin with {...alexPlugin}. The spreading in this case only copies from the top-level and does not copy nested objects, which is problematic in the case of the ESLint plugin which is a very nested structure so we would want to ensure that all nested properties are also copied. We can achieve this deep copying by using the deepCopy function from my utility package.