Abstract out the restricted imports and spread the general ones into the rest of them
I have recently found out that if you have separate no-restricted-imports invocations in different configs, the one listed later always takes priority. As such, for most of these composed configs, the no-restricted-imports from the general JavaScript config did get overwritten by the project-specific ones. As such, we need to ensure that we spread in the general restricted imports into the project-specific restricted imports so that both are accounted for.