Rename the internal config variables
Although the public configs got renamed recently, I had not yet changed the internal config variables so they still ended up using the _Base suffix, which felt a bit redundant. As such, they have been renamed to just be named after more or less what they are named on the plugin itself. That is, the general/typescript config has been named generalTypeScript, and you get the idea for the rest of them.