Only require JSDoc comments for exported utility functions in ESLint plugin
ESLint plugin tends to export and pass around a lot of internal helper functions, which are not necessary to add JSDoc comments to given that they are internal helpers. However, there are still some utility functions that should really have them but are harder to target as they were mixed with some other private utilities in the utility folder. As such, to get around this, I've divided the utilities further into public and private folders, and require-jsdoc is enabled on only the public folder in utility. This feels like a good directory change to make anyway as it makes it clearer exactly which functions are intended to be public and which are internal only.