Do not require destructured objects to have JSDoc comments in components
They never really show up in the editor when using components, so it's not really helpful here. It's actually better to do it on the interface instead, as that's where the usage shows when using React components.

Also note that, for some reason, require-param and check-param-names do the exact same thing for some reason. That is, they both error if I don't annotate destructured props. This is very stupid and I don't know why their philosophy is to have multiple rules that feel very similar in spirit rather than one rule to rule them all, but here we are, I guess.