Make the camelToKebab usage more consistent
Previously, there was a mismatch between compile-time and runtime, where consecutive capitals were being preserved at runtime, but not compile-time. This was fine for personal usage, but now that it is being used in public helpers, it would be best if the compile-time and runtime were more in-sync.

Also, note that I have fixed the conversion of TypeScript and JavaScript so they get replaced starting with the capital letter. This is important otherwise the conversion will be inaccurate (e.g. myTypeScriptProject would become mytypescriptProject and therefore mytypescript-project previously)