Enable the recommended config from eslint-plugin-n in general/javascript
These rules enforce further best practices around syntax and import paths depending on whether you're using CommonJS or ES Modules. I have already gone through and dealt with rules from this config that conflict with other rules and/or give false positives, so it should mostly be fine and not require too much refactoring, but it is best to see what happens in other projects once this goes live and see what edge cases need to be fixed.