Add disableSources: true to typedoc config
This is needed because otherwise when we re-generate the docs, it will always show something different even if the JSDoc comments didn't change. This is because the source it links to would differ every time. I don't really see a huge benefit in linking to the source code anyway (isn't that what docs are there to prevent anyway?), so we can disable this.