Replace all imports from src/index
It's not really the best idea to import from the index, given that it is the entrypoint of the package and it runs the risk of importing partially incompletely resolved functions. Types are generally ok since they're compile-time, but even still, I think it's best to just import from the actual source of the thing being used, especially considering the amount of moving parts throughout the plugin.