Tidy up export space
The Prettier rules has been renamed to Prettier config for consistency

I have also decided to stop exporting all the internal helpers and decided to be more explicit with what we export. Now we only export the public utilities, the plugin itself (including the plugin configs), and the external configs (configs not necessarily ESLint related).

I have also added a test for fixOnCondition after moving that to the public utilities.